% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_package.R
\name{Mann_Whitney_U}
\alias{Mann_Whitney_U}
\title{Mann_Whitney_U}
\usage{
Mann_Whitney_U(x, y = NULL, median_test = 0, paired = F)
}
\arguments{
\item{x}{input vector of the first group. Required}

\item{y}{The second group of variables, optional}

\item{median_test}{with 0 be the defult, compare the values of differences of groups is equal to zero}

\item{paired}{A bollen variable, T means two vector input is a paired sets, the defult is False.}
}
\value{
output of a list containg test statistics (V for single group sign test, W for paired group tests) and its p value
}
\description{
Carry out the rank test
}
