% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_package.R
\name{my_rank}
\alias{my_rank}
\title{my_rank}
\usage{
my_rank(x)
}
\arguments{
\item{x}{a vector of numbers}
}
\value{
A vector with average rank of each element of input x
}
\description{
Simulate the  function 'rank()' in the 'base' package
}
